Make a test utility more generally available
