+++ tests/unit/query_model/test_population_validation.py
@@ -1,8 +1,5 @@
-import dataclasses
-
 import pytest
 
-from databuilder.query_model import nodes as query_model
 from databuilder.query_model.nodes import (
     AggregateByPatient,
     Case,
@@ -20,6 +17,7 @@ from databuilder.query_model.population_validation import (
     evaluate,
     validate_population_definition,
 )
+from tests.lib.query_model_utils import get_all_operations
 
 
 def test_rejects_non_series():
@@ -220,33 +218,6 @@ def test_evaluate(expected, query):
 # Test that the `evaluate()` single dispatch function is exhaustively defined so we
 # can't forget to update it if we add a new query model operation.
 #
-
-
-def get_all_operations():
-    "Return every operation defined in the query model"
-    return [cls for cls in iterate_query_model_namespace() if is_operation(cls)]
-
-
-def is_operation(cls):
-    "Return whether an arbitrary value is a query model operation class"
-    # We need to check this first or the `issubclass` check can fail
-    if not isinstance(cls, type):
-        return False
-    # We need to check it's a proper subclass as the Node base class isn't itself a
-    # dataclass so the `fields()` call will fail
-    if not issubclass(cls, query_model.Node) or cls is query_model.Node:
-        return False
-    # If it takes arguments it's an operation, otherwise it's an abstract type
-    return len(dataclasses.fields(cls)) > 0
-
-
-def iterate_query_model_namespace():
-    "Yield every public thing in the query_model module"
-    yield from [getattr(query_model, name) for name in query_model.__all__]
-    yield from vars(query_model.Function).values()
-    yield from vars(query_model.AggregateByPatient).values()
-
-
 @pytest.mark.parametrize(
     "operation", [op for op in get_all_operations() if issubclass(op, Series)]
 )
